requirements.txt
#to iinstall all: pip install -r requirements_dev.txt
#for testing
# pytest==6.2.4

#core packages for Data analysis
# numpy
pandas
# scipy
# scikit-learn
#ML and DL
# torch
# torchvision
# transformers
openai
# lpips

#UI packages
streamlit
# langdetect
Pillow
# Data visualization
# matplotlib
altair

# API requests and environment variables
requests
python-dotenv
Pillow

#path
Pathlib

#google drive authentication
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
# #analysis
# tabulate

# #development tools
# #pip==19.2.3
# bump2version==0.5.11
# #wheel==0.33.6
# tox==3.14.0
# coverage==4.5.4
# Sphinx==7.2.6
# twine==5.0.0
# ruff==0.3.5
# watchdog>=2.1.0